Move cache classes to own directory
