+++ packages/flame/lib/src/components/cache/memory_cache.dart
