+++ packages/flame/lib/src/components/cache/value_cache.dart
