+++ packages/flame/lib/src/geometry/polygon.dart
@@ -3,10 +3,10 @@ import 'dart:ui' hide Canvas;
 
 import '../../game.dart';
 import '../../geometry.dart';
+import '../components/cache/value_cache.dart';
 import '../extensions/canvas.dart';
 import '../extensions/rect.dart';
 import '../extensions/vector2.dart';
-import '../value_cache.dart';
 import 'shape.dart';
 
 class Polygon extends Shape {
