+++ packages/flame/lib/src/geometry/shape.dart
@@ -3,8 +3,8 @@ import 'dart:ui';
 import '../../components.dart';
 import '../../game.dart';
 import '../../palette.dart';
+import '../components/cache/value_cache.dart';
 import '../extensions/vector2.dart';
-import '../value_cache.dart';
 import 'shape_intersections.dart' as intersection_system;
 
 /// A shape can represent any geometrical shape with optionally a size, position
